requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
mlxtend>=0.22.0
statsmodels>=0.14.0
pyarrow>=12.0.0
pyyaml>=6.0
jupyter>=1.0.0
papermill>=2.4.0
imbalanced-learn>=0.11.0
xgboost>=1.7.0
scipy>=1.11.0
kneed>=0.8.0
